backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
python-dotenv
requests
clerk
cohere